% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mread_rxn.R
\name{mread_rxn}
\alias{mread_rxn}
\title{Load a model file with ODEs specified as reactions}
\usage{
mread_rxn(file, output_file = NULL, ...)
}
\arguments{
\item{file}{full path to file to read, parse, and compile}

\item{output_file}{full path to file to write output}

\item{...}{passed to \link[mrgsolve:mread_cache]{mrgsolve::mread_cache}}
}
\description{
Load a model file with ODEs specified as reactions
}
